# apps/api/openapi/openapi-v1.json
{
...
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/airports" : {
      "get" : {
        "operationId" : "airports",
        "parameters" : [ {
          "in" : "query",
          "name" : "bbox",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "watched",
          "required" : false,
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/geo+json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
    "/api/v1/airports/{icao}/wx" : {
      "get" : {
        "operationId" : "airportWx",
        "parameters" : [ {
          "in" : "path",
          "name" : "icao",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
...
    "/api/v1/alerts" : {
      "get" : {
        "operationId" : "alerts",
        "parameters" : [ {
          "in" : "query",
          "name" : "kind",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
    "/api/v1/alerts/history" : {
      "get" : {
        "operationId" : "alertHistory",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "hex",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
...
    "/api/v1/radar/kr/{tm}.png" : {
      "get" : {
        "operationId" : "radarKrFrame",
        "parameters" : [ {
          "in" : "path",
          "name" : "tm",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
...
    "/api/v1/sigmets" : {
      "get" : {
        "operationId" : "sigmets",
        "parameters" : [ {
          "in" : "query",
          "name" : "active",
          "required" : false,
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "bbox",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "hazard",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/geo+json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
    "/api/v1/sigmets/{id}" : {
      "get" : {
        "operationId" : "sigmet",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "weather-controller" ]
      }
    },
...
  }
}